Moved DMN test
