+++ kie-maven-plugin-example/src/test/java/org/kie/kproject/KProjectTest.java
@@ -15,8 +15,6 @@
 
 package org.kie.kproject;
 
-import java.math.BigDecimal;
-
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -24,20 +22,12 @@ import org.kie.api.KieServices;
 import org.kie.api.runtime.KieContainer;
 import org.kie.api.runtime.KieSession;
 import org.kie.api.runtime.rule.FactHandle;
-import org.kie.dmn.api.core.DMNContext;
-import org.kie.dmn.api.core.DMNModel;
-import org.kie.dmn.api.core.DMNResult;
-import org.kie.dmn.api.core.DMNRuntime;
-import org.kie.dmn.core.api.DMNFactory;
 import org.kie.dmn.core.compiler.ExecModelCompilerOption;
 import org.kie.firealarm.Fire;
 import org.kie.firealarm.Room;
 import org.kie.firealarm.Sprinkler;
 
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.CoreMatchers.notNullValue;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertThat;
+import static org.junit.Assert.*;
 
 public class KProjectTest {
 
@@ -60,24 +50,6 @@ public class KProjectTest {
         kSession.delete(fireFH);
         rules = kSession.fireAllRules();
         assertEquals(3, rules);
-    }
-
-    @Test
-    public void testSolutionCase1() {
-        DMNRuntime runtime = DMNRuntimeUtil.createRuntime("org/kie/example/0020-vacation-days.dmn", this.getClass());
-        DMNModel dmnModel = runtime.getModel("https://www.drools.org/kie-dmn", "0020-vacation-days");
-        assertThat(dmnModel, notNullValue());
-
-        DMNContext context = DMNFactory.newContext();
-
-        context.set("Age", 16);
-        context.set("Years of Service", 1);
-
-        DMNResult dmnResult = runtime.evaluateAll(dmnModel, context);
-
-        DMNContext result = dmnResult.getContext();
-
-        assertThat(result.get("Total Vacation Days"), is(BigDecimal.valueOf(27)));
     }
 
     @Before
