+++ kie-maven-plugin-example/src/test/java/org/kie/kproject/Utils.java
@@ -1,4 +1,4 @@
-package org.acme.test_generateModel_kjararchetype;
+package org.kie.kproject;
 
 import java.math.BigDecimal;
 import java.math.MathContext;
